[追加]FavoriteのSystemSpecを作成した
